Add files via Upload
Uploaded Jupyter Notebook for machine learning project for the Sklearn toy dataset 'diabetes'. The model chosen was GradientBoostingRegressor. Since the model is preprocessed it is suggested in the conclusion to have the original raw data to allow further engineering in hopes to help the modeling.

The documentation for this dataset can be found here: https://scikit-learn.org/stable/modules/generated/sklearn.datasets.load_diabetes.html#sklearn.datasets.load_diabetes